Starting structure
The starting structure to solve the first MED challenge task has been developed. The "hovering node" publishes the pose (x,y,z) that the drone has to reach: a quote of 1 meter from the ground, while x and y reamin equal to zero. Later, the "controller node", by using the data acquired from the aircraft, compute the propellers angular velocities to reach the target. The velocities are published on the specific topic: "command/motor_speed".