Bugs fixed in the waypoint filter
In the waypoint filter, the bugs have been fixed. Now you can enable or disable the filter from the launch file. Furthermore, the filter parameters can be changed via the suitable YAML file.
The angles \theta_r and \phi_r are limited to 30°.
The second and third term in the "calculateRotorsVelocity" function take into account the different length of the arms.